feat(dashboard): move subscriptions to action menu
