+++ frontend/src/metabase/dashboard/components/DashboardHeader/buttons/DashboardActionMenu.tsx
@@ -4,8 +4,12 @@ import type { WithRouterProps } from "react-router/lib/withRouter";
 import { c, t } from "ttag";
 
 import { ToolbarButton } from "metabase/common/components/ToolbarButton";
+import { setSharing as setDashboardSubscriptionSidebarOpen } from "metabase/dashboard/actions";
 import { useDashboardContext } from "metabase/dashboard/context/context";
 import { useRefreshDashboard } from "metabase/dashboard/hooks";
+import { getIsSharing as getIsDashboardSubscriptionSidebarOpen } from "metabase/dashboard/selectors";
+import { useDispatch, useSelector } from "metabase/lib/redux";
+import { DashboardSubscriptionMenuItem } from "metabase/notifications/NotificationsActionsMenu/DashboardSubscriptionMenuItem";
 import { useRegisterShortcut } from "metabase/palette/hooks/useRegisterShortcut";
 import { PLUGIN_MODERATION } from "metabase/plugins";
 import { Icon, Menu } from "metabase/ui";
@@ -37,6 +41,7 @@ const DashboardActionMenuInner = ({
   const { dashboard, isFullscreen, onFullscreenChange, onChangeLocation } =
     useDashboardContext();
   const [opened, setOpened] = useState(false);
+  const dispatch = useDispatch();
 
   const { refreshDashboard } = useRefreshDashboard({
     dashboardId: dashboard?.id ?? null,
@@ -47,6 +52,15 @@ const DashboardActionMenuInner = ({
     dashboard ?? undefined,
     refreshDashboard,
   );
+
+  const isDashboardSubscriptionSidebarOpen = useSelector(
+    getIsDashboardSubscriptionSidebarOpen,
+  );
+
+  const toggleSubscriptionSidebar = () =>
+    dispatch(
+      setDashboardSubscriptionSidebarOpen(!isDashboardSubscriptionSidebarOpen),
+    );
 
   // solely for the dependency list below, so we don't ever have an undefined
   const pathname = location?.pathname ?? "";
@@ -88,6 +102,11 @@ const DashboardActionMenuInner = ({
             {t`Reset all filters`}
           </Menu.Item>
         )}
+
+        <DashboardSubscriptionMenuItem
+          dashboard={dashboard}
+          onClick={toggleSubscriptionSidebar}
+        />
 
         <Menu.Item
           leftSection={<Icon name="expand" />}
